refactor(test-consume): share the multi-test client fixtures
The enginex and wirex conftests carried near-verbatim copies of the
client lifecycle plumbing: the client fixture, the genesis_header
fixture, the tracker wiring and the per-test Hive reporting hook,
plus the group-counting half of the collection hook. The copies had
already drifted in log lines, and any fix to the client lifecycle
would have had to land twice.

Move the four fixtures into multi_test_client, which both simulators
already load and which owns the rest of the group-scoped plumbing
(pre_alloc_group, client_genesis, environment), and extract the group
counting into a helper next to the stash key it fills. Each
simulator's conftest keeps only its own collection policy: enginex
sorts largest group first, wirex additionally orders the tests inside
each group.
