+++ packages/testing/src/execution_testing/cli/pytest_commands/plugins/consume/simulators/helpers/test_tracker.py
@@ -13,7 +13,38 @@ logger = logging.getLogger(__name__)
 enginex_group_counts_key: StashKey[dict[str, int]] = StashKey()
 
 
-def make_group_identifier(pre_hash: AllocGroupHash, client_name: str) -> str:
+def count_tests_per_group(
+    session: pytest.Session, items: list[pytest.Item]
+) -> dict[str, int]:
+    """
+    Count the collected tests of each pre-allocation group and stash
+    the counts on the session.
+
+    The xdist_group markers are set during parametrization. The counts
+    feed the largest-group-first sort of the simulators' collection
+    hooks and, via the session stash, the test tracker's client
+    teardown accounting.
+    """
+    group_counts: dict[str, int] = {}
+    for item in items:
+        for marker in item.iter_markers("xdist_group"):
+            if "name" in marker.kwargs:
+                group_counts[marker.kwargs["name"]] = (
+                    group_counts.get(marker.kwargs["name"], 0) + 1
+                )
+                break
+
+    session.stash[enginex_group_counts_key] = group_counts
+    logger.info(
+        f"Counted {len(group_counts)} pre-alloc groups with "
+        f"{sum(group_counts.values())} total tests"
+    )
+    return group_counts
+
+
+def make_group_identifier(
+    pre_hash: AllocGroupHash, client_name: str
+) -> str:
     """Build xdist group key from pre-alloc hash and client name."""
     return f"{pre_hash}-{client_name}"
